+++ karma.conf.js
@@ -1,6 +1,0 @@
-/** @param {import('karma').Config} config */
-module.exports = config => config.set({
-  client: {
-    captureConsole: false
-  },
-});