+++ src/components/PatronBlock/PatronBlockForm.test.js
@@ -4,7 +4,9 @@ import {
   screen,
   cleanup,
 } from '@testing-library/react';
+import { within } from '@testing-library/dom';
 import userEvent from '@testing-library/user-event';
+
 import renderWithRouter from 'helpers/renderWithRouter';
 import PatronBlockForm from './PatronBlockForm';
 import '__mock__/matchMedia.mock';
@@ -63,18 +65,18 @@ const props = {
   onSubmit: jest.fn(),
   connect: jest.fn(),
   intl: {
-    formatMessage: jest.fn(),
+    formatMessage: ({ id }) => id,
   },
   stripes: STRIPES,
   initialValues: {},
   blockTemplates: [
     {
-      id: 1,
+      id: '1',
       code: 'testCode1',
       name: 'name1'
     },
     {
-      id: 2,
+      id: '2',
       code: 'testCode2',
       name: 'name2'
     }
@@ -95,17 +97,27 @@ describe('Patron Block Form', () => {
   it('should render the component', () => {
     expect(screen.getByText('ui-users.blocks.form.button.create')).toBeInTheDocument();
   });
+
   it('Toggle button must work', () => {
     userEvent.click(document.querySelector('[id=accordion-toggle-button-blockInformationSection]'));
     expect(screen.getByText('ui-users.blocks.form.label.information')).toBeInTheDocument();
   });
+
   it('Expand button must work', () => {
-    userEvent.click(document.querySelector('[data-tast-expand-button="true"]'));
+    userEvent.click(document.querySelector('[data-test-expand-button="true"]'));
     expect(screen.getByText('ui-users.blocks.form.label.block')).toBeInTheDocument();
   });
+
   it('Change template must work', () => {
-    userEvent.click(document.querySelector('[id="option-stripes-selection-45-1-1"]'));
-    expect(screen.getByText('name2 (testCode2)')).toBeInTheDocument();
+    // click to open the Selection, then click a specific option.
+    // after clicking, the way stripes-components renders a <Selection>
+    // with a selected element, the selected element will be present twice.
+    userEvent.click(screen.getByText('ui-users.blocks.form.label.template'));
+
+    const list = screen.getByRole('listbox');
+    userEvent.click(within(list).getByText('name2', { exact: false }));
+
+    expect(screen.getAllByText('name2 (testCode2)').length).toEqual(2);
   });
 });