+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change history for ui-users
 
-## 8.3.0 IN PROGRESS
+## 9.0.0 IN PROGRESS
 * Implement modal windows for BILL ACTUAL COST and DO NOT BILL. Refs UIU-2714.
 * create Jest/RTL test for UserDetail.js. Refs UIU-2421
 * Fix problem with displaying suspended claim returned fees/fines. Refs UIU-2726.
@@ -17,6 +17,14 @@
 * In Settings > Users > Patron Blocks > Conditions, mark "Message to be displayed" as required. Refs UIU-2487.
 * Enable dependabot. Refs UIU-2747, FOLIO-3664.
 * Suppress delete of users stored in a configuration entry. Refs UIU-2738.
+* Remove BigTest infrastructure including tests, deps, config. Refs UIU-2745.
+* Add support for `request-storage` version `5.0` for `<UserRequests>`. Refs UIU-2765.
+* Implement "Do not bill actual cost" functionality. Refs UIU-2705.
+* *BREAKING* Upgrade to `@folio/stripes` `v8`. Refs UIU-2761.
+* *BREAKING* Upgrade `react-redux` to `v8`. Refs UIU-2775.
+* Create Jest/RTL test for UserDetailFullscreen.js. Refs UIU-2429.
+* Add support for checkbox custom field filter. Fixes UIU-2759.
+* Use `updateUser` from `@folio/stripes-core` to update service-point data. Refs UIU-2788.
 
 ## [8.2.0](https://github.com/folio-org/ui-users/tree/v8.2.0) (2022-10-24)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v8.1.0...v8.2.0)